moved ConnectivityMode
